% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/somatemplate-package.R
\docType{package}
\name{somatemplate}
\alias{somatemplate}
\alias{somatemplate-package}
\title{somatemplate}
\description{
A template for pkgdown websites developed by SomaLogic Operating Co., Inc.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://somalogic.github.io/somatemplate/}
  \item \url{https://somalogic.com}
  \item Report bugs at \url{https://github.com/SomaLogic/somatemplate/issues}
}

}
\author{
Amanda Hiser
}
\keyword{internal}
